Fix signedness warning in gio/tests/socket.c
gio/tests/socket.c: In function ‘test_get_available’:
gio/tests/socket.c:1661:53: error: comparison of integer expressions of different signedness: ‘gssize’ {aka ‘long int’} and ‘long unsigned int’ [-Werror=sign-compare]
           if (g_socket_get_available_bytes (server) > sizeof (data))
                                                     ^

+++ gio/tests/socket.c
@@ -1658,7 +1658,7 @@ test_get_available (gconstpointer user_data)
 
       for (tries = 0; tries < 100; tries++)
         {
-          if (g_socket_get_available_bytes (server) > sizeof (data))
+          if ((gsize) g_socket_get_available_bytes (server) > sizeof (data))
             break;
           g_usleep (100000);
         }
